requirements.txt
# Core LangChain + LangGraph
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.1.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0

# PostgreSQL + pgvector (replaces chromadb + rank-bm25)
psycopg2-binary>=2.9.9
pgvector>=0.3.0
langchain-postgres>=0.0.9

# RAG + document processing
pypdf>=4.0.0
python-docx>=1.1.0
rapidfuzz>=3.6.0

# OCR support (optional — CPU; swap paddlepaddle → paddlepaddle-gpu for CUDA)
paddlepaddle>=2.6.0
paddleocr>=2.7.0
pymupdf>=1.24.0
pillow>=10.0.0

# Observability
langfuse>=2.0.0

# CLI / utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typer>=0.12.0
rich>=13.0.0
